Tuesday

+++ main.py
@@ -5,6 +5,9 @@ import gptmodules
 import sqlalchemy
 from databricks import sql
 from credentials import dbxcat, dbxhost, dbxpath, dbxschema, dbxtoken
+import os
+import openpyxl
+
 
 
 def main():
@@ -20,6 +23,9 @@ def main():
 
     no_of_files = len(files)
     #no_of_files = 1
+
+    
+    
     finished_data = pd.DataFrame()
     provided_data = pd.DataFrame()
     order_item = pd.DataFrame()
@@ -27,50 +33,41 @@ def main():
     # ======= start iteration here ========= 
     for i in range(no_of_files):
         filename = files[i]['name']
-        file_id = modules.get_file_id(access_token, site_id, drive_id, filename)
-        modules.download_file(access_token, site_id, drive_id, file_id, filename)
+        if str('.pdf') in str(filename):
+                      
+            file_id = modules.get_file_id(access_token, site_id, drive_id, filename)
+            modules.download_file(access_token, site_id, drive_id, file_id, filename)
 
-        # scan the file - finished data 
-        f_data = gptmodules.get_finished_data(filename, credentials.GPT4V_KEY, credentials.GPT4V_ENDPOINT)
-        finished_data = pd.concat([finished_data, f_data], ignore_index=True)
-        #print(f_data)
+            # scan the file - finished data 
+            f_data = gptmodules.get_finished_data(filename, credentials.GPT4V_KEY, credentials.GPT4V_ENDPOINT)
+            finished_data = pd.concat([finished_data, f_data], ignore_index=True)
+            #print(f_data)
 
-        # scan the file - provided material
-        p_data = gptmodules.get_provided_data(filename, credentials.GPT4V_KEY, credentials.GPT4V_ENDPOINT)
-        provided_data = pd.concat([provided_data, p_data], ignore_index=True)
-        #print(p_data)
+            # scan the file - provided material
+            p_data = gptmodules.get_provided_data(filename, credentials.GPT4V_KEY, credentials.GPT4V_ENDPOINT)
+            provided_data = pd.concat([provided_data, p_data], ignore_index=True)
+            #print(p_data)
         
-        # scan the filename - order item
-        o_item = modules.extract_numbers(filename)
-        order_item = o_item[0]
-        print(order_item)
+            # scan the filename - order item
+            o_item = modules.extract_numbers(filename)
+            order_item = o_item[0]
+            print(order_item)
 
-        # delete the pdf file and pngs
-        modules.delete_file(filename)
-        f_data['Material No'] = str(order_item)
-        merged_df = modules.alternate_rows(f_data, p_data)
-        merged_df['Remarks'] = " "
-        print(merged_df.to_markdown())
-        fname = "GVT2-" + order_item + ".xlsx"
-        merged_df.to_excel(fname, index=False, sheet_name=order_item)
-        
-    
-    #finished_data.to_excel('finished_product.xlsx', index=False)
-    #provided_data.to_excel('provided_materials.xlsx', index=False)
+            # delete the pdf file and pngs
+            modules.delete_file(filename)
+            f_data['Material No'] = str(order_item)
+            merged_df = modules.alternate_rows(f_data, p_data)
+            merged_df['Remarks'] = " "
+            #print(merged_df.to_markdown())
+            fname = "GVT2-" + order_item + ".xlsx"
+            merged_df.to_excel(fname, index=False, sheet_name=order_item)
 
     # ===== Databricks =====
-    dbxcat = "iaut_dmart"
-    dbxschema = "exp_ic_162105_open_items"
-    dbxhost = "adb-4071335540424391.11.azuredatabricks.net"
-    dbxpath = "/sql/1.0/warehouses/63b06d5d61cc1af1"
-    dbxtoken = "dapi1d3596c872e65f3f95430ff5a75f6d90-2"
-
-    connection = sql.connect(server_hostname = dbxhost, http_path = dbxpath, access_token = dbxtoken)
-    cursor = connection.cursor()
-
-    filename = "exp_ic_162105_open_items"
-    local_path = filename + ".xlsx"
-    flag = 0
+    dbxcat = "efdataonelh_prd"
+    dbxschema = "iaut_dmart"
+    dbxhost     = os.environ["DBXHOST"]                                       ## Credentials
+    dbxpath     = os.environ["DBXPATH"]                                       ## Credentials
+    dbxtoken    = os.environ["DBXTOKEN"]          
 
     extra_connect_args = {
         "_tls_verify_hostname": True,
@@ -84,23 +81,72 @@ def main():
     )
 
     cnxn = engine.connect()
+    # === EEND Databricks
 
-    pvd = pd.DataFrame()
-    for i in range(len(provided_data)):
-        mn = str(provided_data['Material No'][i]).zfill(len(str(provided_data['Material No'][i]))+10)
-        #print(mn)
-        bn = provided_data['Batch'][i]
-        #bn = finished_data[i]['Batch No']
-        sqlquery = "SELECT MATNR, CHARG, CINSM, LFMON FROM efdataonelh_prd.generaldiscovery_matmgt_r.all_mchbh_view where MATNR = '" + mn + "' AND CHARG = '" + bn + "'  AND LFGJA = '2024'"
-        #print(sqlquery)
-        try:
-            df = pd.read_sql_query(sqlquery, cnxn)
-            pvd = pd.concat([pvd, df], ignore_index=True)
+    # list all files in the directory
+    files = os.listdir()
+
+    # Convert the list of files to a DataFrame
+    df = pd.DataFrame(files, columns=['Filename'])
+    
+    # filter out only GVT2 Excel files
+    gvt2_df = df[df['Filename'].str.contains('GVT2-')]
+  
+    
+    # loop around the numbers of GVT2 excel files
+    for i in range(len(gvt2_df)):
+        filename = gvt2_df.iloc[i]['Filename']
+        print(filename)
+        
+        # load the excel file to a temporary dataframe
+        df = pd.read_excel(filename)
+        print(df.to_markdown())
+        
+        #loop around the df:
+        for j in range(len(df)):
+            type = df.iloc[j]['Material Type']
+            matno = str(df.iloc[j]['Material No'])
+            batch = str(df.iloc[j]['Batch'])
+            qty = df.iloc[j]['Qty']
+            #print(type)
             
-        except:
-            print("error")
-
-    #print(pvd.to_markdown())
-
+            # IF type is Bulk
+            if type == 'Bulk':
+                #print('bulk')
+                sqlquery = "SELECT MATNR, CHARG, CINSM, CLABS, LFMON, LFGJA FROM efdataonelh_prd.generaldiscovery_matmgt_r.all_mchbh_view where MATNR LIKE '%" + matno + "' AND CHARG = '" + batch + "' AND LFGJA IN ('2024','2025')"
+                #print(sqlquery)
+                try:
+                    df2 = pd.read_sql_query(sqlquery, cnxn)
+                    dbx_qty = (df2['CLABS'].sum()) * 1000
+                    # write the quantity to the excel file
+                    workbook = openpyxl.load_workbook(filename=filename)
+                    sheet = workbook.active
+                    cellno = "E" + str(j+2)
+                    sheet[cellno] = dbx_qty 
+                    workbook.save(filename)
+                    #print(dbx_qty)
+                    #print(df2.to_markdown())
+            
+                except:
+                    print("error")
+            else:
+                sqlquery = "SELECT MATNR, CHARG, CINSM, CLABS, LFMON, LFGJA FROM efdataonelh_prd.generaldiscovery_matmgt_r.all_mchbh_view where MATNR LIKE '%" + matno + "' AND CHARG = '" + batch + "' AND LFGJA IN ('2024','2025')"
+                #print(sqlquery)
+                try:
+                    df2 = pd.read_sql_query(sqlquery, cnxn)
+                    dbx_qty = df2['CINSM'].sum()
+                    # write the quantity to the excel file
+                    workbook = openpyxl.load_workbook(filename=filename)
+                    sheet = workbook.active
+                    cellno = "E" + str(j+2)
+                    sheet[cellno] = dbx_qty
+                    workbook.save(filename)
+                    #print(dbx_qty)
+                    #print(df2.to_markdown())
+            
+                except:
+                    print("error")
+            
+    
 if __name__ == "__main__":
     main()
